webserver

+++ test/test_webserver.py
@@ -1,20 +1,33 @@
-def test_webserver():
+start_server = True
+#start_server = False
 
-    # initiate server
-    print('> Starting server')
+def test_webserver():
 
     import subprocess
     import time
     from pyctrl.util.json import JSONDecoder, JSONEncoder
-    
-    server = subprocess.Popen(["python3",
-                               "pyctrl/webserver.py"],
-                              stdout = subprocess.PIPE)
+        
+    if start_server:
+        
+        # initiate server
+        print('> Starting server')
 
-    time.sleep(1)
+        server = subprocess.Popen(["python3",
+                                   "pyctrl/webserver.py"],
+                                  stdout = subprocess.PIPE)
+        
+        time.sleep(1)
 
     try:
 
+        # reset controller
+        url = "http://127.0.0.1:5000/reset"
+        output = subprocess.check_output(["curl", url]).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'status': 'success'}
+        
+        assert result == answer
+        
         # check index page
         answer = b"<div><p>&lt;class 'pyctrl.Controller'&gt; with: 0 timer(s), 3 signal(s), 1 source(s), 0 filter(s), and 0 sink(s)</p><h2>timers</h2><ol></ol><h2>signals</h2><ol><li>clock</li><li>duty</li><li>is_running</li></ol><h2>sources</h2><ol><li>clock[Clock, disabled] &Gt; clock</li></ol><h2>filters</h2><ol></ol><h2>sinks</h2><ol></ol></div>"
 
@@ -27,7 +40,7 @@ def test_webserver():
         output = subprocess.check_output(["curl", url])
         assert output == answer
 
-        # sink methods
+        # S I N K S
         
         # add sink
         url = r'"http://127.0.0.1:5000/add/sink/printer/pyctrl.block/Printer?inputs=\[\"clock\",\"is_running\"\]"'
@@ -109,6 +122,276 @@ def test_webserver():
 
         assert result == answer
         
+        # add sink with parameter
+        url = r'"http://127.0.0.1:5000/add/sink/printer/pyctrl.block/Printer?inputs=\[\"clock\",\"is_running\"\]&kwargs=\{\"endln\":\"\\r\"\}"'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'status': 'success'}
+
+        assert result == answer
+
+        # get attribute/sink (multiple)
+        url = '"http://127.0.0.1:5000/get/sink/printer?keys=\[\\"enabled\\",\\"endln\\"\]"'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'enabled': True, 'endln': '\r' }
+
+        assert result == answer
+        
+        # S O U R C E
+        
+        # add source
+        url = r'"http://127.0.0.1:5000/add/source/constant/pyctrl.block/Constant?outputs=\[\"signal\"\]&kwargs=\{\"value\":3\}"'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'status': 'success'}
+
+        assert result == answer
+
+        # get source
+        url = "http://127.0.0.1:5000/get/source/constant"
+        output = subprocess.check_output(["curl", url]).decode("utf-8")
+        result = JSONDecoder().decode(output)
+
+        from pyctrl.block import Constant
+        assert result['constant'].get() == Constant(value = 3).get()
+
+        # get attribute/source
+        url = r'"http://127.0.0.1:5000/get/source/constant?keys=\"enabled\""'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'enabled': True}
+
+        assert result == answer
+
+        # get attribute/source (multiple)
+        url = r'"http://127.0.0.1:5000/get/source/constant?keys=\[\"enabled\",\"value\"\]"'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'enabled': True, 'value': 3}
+
+        assert result == answer
+        
+        # set attribute/source
+        url = 'http://127.0.0.1:5000/set/source/constant?enabled=false'
+        output = subprocess.check_output(["curl", url]).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'status': 'success'}
+
+        assert result == answer
+
+        # get attribute/source
+        url = '"http://127.0.0.1:5000/get/source/constant?keys=\\"enabled\\""'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'enabled': False}
+
+        assert result == answer
+        
+        # set attribute/source
+        url = '"http://127.0.0.1:5000/set/source/constant?value=4"'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'status': 'success'}
+
+        assert result == answer
+
+        # get attribute/source
+        url = '"http://127.0.0.1:5000/get/source/constant?keys=\\"value\\""'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'value': 4}
+
+        assert result == answer
+        
+        # set attribute/source (multiple)
+        url = '"http://127.0.0.1:5000/set/source/constant?enabled=true&value=5"'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'status': 'success'}
+
+        assert result == answer
+
+        # get attribute/source (multiple)
+        url = '"http://127.0.0.1:5000/get/source/constant?keys=\[\\"enabled\\",\\"value\\"\]"'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'enabled': True, 'value': 5 }
+
+        assert result == answer
+
+        # F I L T E R
+        
+        # add filter
+        url = r'"http://127.0.0.1:5000/add/filter/gain/pyctrl.block.system/Gain?inputs=\[\"inp\"\]&outputs=\[\"out\"\]&kwargs=\{\"gain\":3\}"'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'status': 'success'}
+
+        assert result == answer
+
+        # get filter
+        url = "http://127.0.0.1:5000/get/filter/gain"
+        output = subprocess.check_output(["curl", url]).decode("utf-8")
+        result = JSONDecoder().decode(output)
+
+        from pyctrl.block.system import Gain
+        assert result['gain'].get() == Gain(gain = 3).get()
+
+        # get attribute/filter
+        url = r'"http://127.0.0.1:5000/get/filter/gain?keys=\"enabled\""'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'enabled': True}
+
+        assert result == answer
+
+        # get attribute/filter (multiple)
+        url = r'"http://127.0.0.1:5000/get/filter/gain?keys=\[\"enabled\",\"gain\"\]"'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'enabled': True, 'gain': 3}
+
+        assert result == answer
+        
+        # set attribute/filter
+        url = 'http://127.0.0.1:5000/set/filter/gain?enabled=false'
+        output = subprocess.check_output(["curl", url]).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'status': 'success'}
+
+        assert result == answer
+
+        # get attribute/filter
+        url = '"http://127.0.0.1:5000/get/filter/gain?keys=\\"enabled\\""'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'enabled': False}
+
+        assert result == answer
+        
+        # set attribute/filter
+        url = '"http://127.0.0.1:5000/set/filter/gain?gain=4"'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'status': 'success'}
+
+        assert result == answer
+
+        # get attribute/filter
+        url = '"http://127.0.0.1:5000/get/filter/gain?keys=\\"gain\\""'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'gain': 4}
+
+        assert result == answer
+        
+        # set attribute/filter (multiple)
+        url = '"http://127.0.0.1:5000/set/filter/gain?enabled=true&gain=5"'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'status': 'success'}
+
+        assert result == answer
+
+        # get attribute/filter (multiple)
+        url = '"http://127.0.0.1:5000/get/filter/gain?keys=\[\\"enabled\\",\\"gain\\"\]"'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'enabled': True, 'gain': 5 }
+
+        assert result == answer
+        
+        # T I M E R
+        
+        # add timer
+        url = r'"http://127.0.0.1:5000/add/timer/gain/pyctrl.block.system/Gain?inputs=\[\"inp\"\]&outputs=\[\"out\"\]&kwargs=\{\"gain\":3\}&period=1"'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'status': 'success'}
+
+        assert result == answer
+
+        # get timer
+        url = "http://127.0.0.1:5000/get/timer/gain"
+        output = subprocess.check_output(["curl", url]).decode("utf-8")
+        result = JSONDecoder().decode(output)
+
+        from pyctrl.block.system import Gain
+        assert result['gain'].get() == Gain(gain = 3).get()
+
+        # get attribute/timer
+        url = r'"http://127.0.0.1:5000/get/timer/gain?keys=\"enabled\""'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'enabled': True}
+
+        assert result == answer
+
+        # get attribute/timer (multiple)
+        url = r'"http://127.0.0.1:5000/get/timer/gain?keys=\[\"enabled\",\"gain\"\]"'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'enabled': True, 'gain': 3}
+
+        assert result == answer
+        
+        # set attribute/timer
+        url = 'http://127.0.0.1:5000/set/timer/gain?enabled=false'
+        output = subprocess.check_output(["curl", url]).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'status': 'success'}
+
+        assert result == answer
+
+        # get attribute/timer
+        url = '"http://127.0.0.1:5000/get/timer/gain?keys=\\"enabled\\""'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'enabled': False}
+
+        assert result == answer
+        
+        # set attribute/timer
+        url = '"http://127.0.0.1:5000/set/timer/gain?gain=4"'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'status': 'success'}
+
+        assert result == answer
+
+        # get attribute/timer
+        url = '"http://127.0.0.1:5000/get/timer/gain?keys=\\"gain\\""'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'gain': 4}
+
+        assert result == answer
+        
+        # set attribute/timer (multiple)
+        url = '"http://127.0.0.1:5000/set/timer/gain?enabled=true&gain=5"'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'status': 'success'}
+
+        assert result == answer
+
+        # get attribute/timer (multiple)
+        url = '"http://127.0.0.1:5000/get/timer/gain?keys=\[\\"enabled\\",\\"gain\\"\]"'
+        output = subprocess.check_output('curl ' + url, shell=True).decode("utf-8")
+        result = JSONDecoder().decode(output)
+        answer = {'enabled': True, 'gain': 5 }
+
+        assert result == answer
+
+        answer = b"<div><p>&lt;class 'pyctrl.Controller'&gt; with: 1 timer(s), 6 signal(s), 2 source(s), 1 filter(s), and 1 sink(s)</p><h2>timers</h2><ol><li>inp &Gt; gain[Gain, period = 1, repeat, enabled] &Gt; out</li></ol><h2>signals</h2><ol><li>clock</li><li>duty</li><li>inp</li><li>is_running</li><li>out</li><li>signal</li></ol><h2>sources</h2><ol><li>clock[Clock, disabled] &Gt; clock</li><li>constant[Constant, enabled] &Gt; signal</li></ol><h2>filters</h2><ol><li>inp &Gt; gain[Gain, enabled] &Gt; out</li></ol><h2>sinks</h2><ol><li>clock, is_running &Gt; printer[Printer, enabled]</li></ol></div>"
+        
+        # check info page
+        url = "http://127.0.0.1:5000/info"
+        output = subprocess.check_output(["curl", url])
+
+        assert output == answer
+        
     except Exception as e:
         
         print('** EXCEPTION RAISED **')
@@ -117,6 +400,8 @@ def test_webserver():
     
     finally:
         
-        # stop server
-        print('> Terminating server')
-        server.terminate()
+        if start_server:
+            
+            # stop server
+            print('> Terminating server')
+            server.terminate()
